Fetch all gene set collections

Fetch information about all gene set collections in the Gesel database.

fetchAllCollections(species, config = NULL)

Arguments

species

String containing the NCBI taxonomy ID of the species of interest.

config

Configuration list, typically created by newConfig. If NULL, the default configuration is used.

Value

Data frame of gene set collection information. Each row represents a collection and contains:

title, string containing the title of the collection.

description, string containing a description of the collection.

maintainer, string containing the identity of the collection's maintainer.

source, string containing the source of origin of the collection.

start, integer containing the set index of the first gene set in this collection. The set index refers to a row in the data frame returned by fetchAllSets.

size, integer specifying the number of gene sets in the collection.

Details

If this function is called once, the data frame will be cached in memory and re-used in subsequent calls to this function. The cached data will also be used to speed up calls to fetchSomeCollections.

Author

Aaron Lun

Examples

out <- fetchAllCollections("9606")
head(out)
#>                                                                       title
#> 1                                                             Gene ontology
#> 2                                           MSigDB C1: positional gene sets
#> 3     MSigDB C2: curated gene sets, CGP: chemical and genetic perturbations
#> 4          MSigDB C2: curated gene sets, CP: Canonical pathways, PID subset
#> 5     MSigDB C2: curated gene sets, CP: Canonical pathways, Reactome subset
#> 6 MSigDB C2: curated gene sets, CP: Canonical pathways, WikiPathways subset
#>                                                                                                                                                                                                                                       description
#> 1                                                                                                                           Gene sets defined from the Gene Ontology (version 2026-01-23), sourced from AH121953 in Bioconductor's AnnotationHub.
#> 2                                                                                                                                                           Gene sets corresponding to human chromosome cytogenetic bands. Part of MSigDB 2026.1.
#> 3 Gene sets represent expression signatures of genetic and chemical perturbations. A number of these gene sets come in pairs: xxx_UP (and xxx_DN) gene set representing genes induced (and repressed) by the perturbation. Part of MSigDB 2026.1.
#> 4                                                                                                                                                      Canonical Pathways gene sets derived from the PID pathway database. Part of MSigDB 2026.1.
#> 5                                                                                                                                                 Canonical Pathways gene sets derived from the Reactome pathway database. Part of MSigDB 2026.1.
#> 6                                                                                                                                             Canonical Pathways gene sets derived from the WikiPathways pathway database. Part of MSigDB 2026.1.
#>   maintainer
#> 1  Aaron Lun
#> 2  Aaron Lun
#> 3  Aaron Lun
#> 4  Aaron Lun
#> 5  Aaron Lun
#> 6  Aaron Lun
#>                                                                     source
#> 1         https://github.com/gesel-inc/feedstock/blob/go-v1.1.0/go/build.R
#> 2 https://github.com/gesel-inc/feedstock/blob/msigdb-v1.1.0/msigdb/build.R
#> 3 https://github.com/gesel-inc/feedstock/blob/msigdb-v1.1.0/msigdb/build.R
#> 4 https://github.com/gesel-inc/feedstock/blob/msigdb-v1.1.0/msigdb/build.R
#> 5 https://github.com/gesel-inc/feedstock/blob/msigdb-v1.1.0/msigdb/build.R
#> 6 https://github.com/gesel-inc/feedstock/blob/msigdb-v1.1.0/msigdb/build.R
#>   start  size
#> 1     1 18864
#> 2 18865   302
#> 3 19167  3555
#> 4 22722   196
#> 5 22918  1839
#> 6 24757   925