Fetch sets for all genes

Fetch the identities of the sets that contain each gene in the Gesel database.

fetchSetsForAllGenes(
  species,
  fetch = downloadDatabaseFile,
  fetch.args = list()
)

Arguments

species

String containing the NCBI taxonomy ID of the species of interest.

fetch

Function that accepts the name of a Gesel database file and returns an absolute path to that file.

fetch.args

Named list of arguments to pass to fetch.

Value

List of integer vectors. Each vector corresponds to a gene in the same order as fetchAllGenes. Each vector contains the identities of the sets that contain that gene, where each integer is a set index that refers to a row of the data frame returned by fetchAllSets.

Author

Aaron Lun

Examples

all.genes <- fetchSetsForAllGenes("9606")
length(all.genes)
#> [1] 105830

# Sets containing the first gene:
head(fetchAllSets("9606")[all.genes[[1]],])
#>            name                  description size collection number
#> 1327 GO:0003674           molecular_function  710          1   1327
#> 2337 GO:0005576         extracellular region 1916          1   2337
#> 2366 GO:0005615          extracellular space 1865          1   2366
#> 3538 GO:0008150           biological_process  561          1   3538
#> 6639 GO:0031093 platelet alpha granule lumen   67          1   6639
#> 8166 GO:0034774      secretory granule lumen  115          1   8166

# Details about the first gene:
fetchAllGenes("9606")$symbol[1]
#> [[1]]
#> [1] "A1BG"
#>