Load all gene sets for a species

Load information about all gene sets for a species. This includes its gene membership as well as details like its name and description.

loadAllSets(species, type, config = NULL, as.compressed = FALSE)

Arguments

species

String containing the NCBI taxonomy ID of the species of interest.

type

String specifying the type of gene identifier to report. This can be any type listed in fetchGeneTypes.

config

Configuration list, typically created by newConfig. If NULL, the default configuration is used.

as.compressed

Boolean indicating whether to return a CompressedCharacterList.

Value

If as.compressed = FALSE, a list is returned containing:

sets, a list of length equal to the total number of sets for species. Each element is a character vector that corresponds to a gene set and contains the genes in that set. Each gene is represented by zero, one or more identifiers of the specified type (see renameGenesInSets for details). The positional index of each set in sets can be used as the Gesel set index in other gesel functions like fetchGenesForSomeSets.

details, a data frame where each row corresponds to a gene set in sets. The data frame contains the following columns:

name, string containing the name of the gene set.

description, string containing a description of the gene set.

size, integer specifying the number of genes in this gene set. Note that this may not equal lengths(sets) if any gene does not have exactly one identifier of the specified type.

collection, the name of the collection that contains this gene set.

The row index of each set in details can be used as the Gesel set index in other gesel functions like fetchGenesForSomeSets.

If as.compressed = TRUE, a CompressedCharacterList of length equal to the total number of sets for species is returned. Each element is a character vector that contains the identifiers of genes in that set, as described above for sets. The mcols contains more details about each set as described above for details.

See also

Author

Aaron Lun

Examples

everything <- loadAllSets("7227", "symbol")
head(everything$sets)
#> [[1]]
#> [1] "Pink1" "larp" 
#> 
#> [[2]]
#> [1] "Alg12" "PIG-V"
#> 
#> [[3]]
#> [1] "XRCC1"  "CG6171" "Aptx"   "CG5316"
#> 
#> [[4]]
#>  [1] "mei-9"   "Tengl3"  "mre11"   "CG12917" "EndoG"   "CG14120" "CG14118"
#>  [8] "CG6839"  "CG3819"  "Tengl4"  "CG14062" "Sid"     "Tengl2"  "Tengl1" 
#> [15] "CG33346"
#> 
#> [[5]]
#> [1] "Eno"
#> 
#> [[6]]
#> [1] "Eb1"   "Mgtor" "Mtor"  "msps"  "ncd"   "chb"   "Rok"  
#> 
head(everything$details)
#>         name                                        description size
#> 1 GO:0000001                          mitochondrion inheritance    2
#> 2 GO:0000009             alpha-1,6-mannosyltransferase activity    2
#> 3 GO:0000012                         single strand break repair    3
#> 4 GO:0000014 single-stranded DNA endodeoxyribonuclease activity   15
#> 5 GO:0000015                  phosphopyruvate hydratase complex    1
#> 6 GO:0000022                         mitotic spindle elongation    6
#>      collection
#> 1 Gene ontology
#> 2 Gene ontology
#> 3 Gene ontology
#> 4 Gene ontology
#> 5 Gene ontology
#> 6 Gene ontology

everything2 <- loadAllSets("7227", "symbol", as.compressed = TRUE)
everything2
#> CharacterList of length 8707
#> [[1]] Pink1 larp
#> [[2]] Alg12 PIG-V
#> [[3]] XRCC1 CG6171 Aptx CG5316
#> [[4]] mei-9 Tengl3 mre11 CG12917 EndoG ... CG14062 Sid Tengl2 Tengl1 CG33346
#> [[5]] Eno
#> [[6]] Eb1 Mgtor Mtor msps ncd chb Rok
#> [[7]] PIG-B Alg9 PIG-Z
#> [[8]] CG1785 RpLP0-like CG13185 Non3 CG11583 RpL5
#> [[9]] sta RpS28b RpS19a CG9107 CG12325 ... RpS28a RpS14a RpS14b RpS28-like
#> [[10]] tw CG6659 Tmtc2 CG4341 Tmtc1 CG31690 ... Alg3 rt Tmtc4 CG5038 Alg1 PIG-V
#> ...
#> <8697 more elements>
S4Vectors::mcols(everything2)
#> DataFrame with 8707 rows and 4 columns
#>             name            description      size    collection
#>      <character>            <character> <integer>   <character>
#> 1     GO:0000001 mitochondrion inheri..         2 Gene ontology
#> 2     GO:0000009 alpha-1,6-mannosyltr..         2 Gene ontology
#> 3     GO:0000012 single strand break ..         3 Gene ontology
#> 4     GO:0000014 single-stranded DNA ..        15 Gene ontology
#> 5     GO:0000015 phosphopyruvate hydr..         1 Gene ontology
#> ...          ...                    ...       ...           ...
#> 8703  GO:2001256 regulation of store-..         1 Gene ontology
#> 8704  GO:2001261 negative regulation ..         1 Gene ontology
#> 8705  GO:2001262 positive regulation ..         1 Gene ontology
#> 8706  GO:2001282 negative regulation ..         1 Gene ontology
#> 8707  GO:7770001 mitochondrial pyruva..         3 Gene ontology