Search set text

Search for sets based on their names and descriptions.

searchSetText(
  species,
  query,
  use.name = TRUE,
  use.description = TRUE,
  config = NULL
)

Arguments

species

String containing the NCBI taxonomy ID of the species of interest.

query

String containing one or more words to search on. A set is only matched if it matches to all of the tokens in the query. The * and ? wildcards can be used to match to any or one character, respectively.

use.name

Logical scalar indicating whether to search on the name of the set.

use.description

Logical scalar indicating whether to search on the description of the set.

config

Configuration list, typically created by newConfig. If NULL, the default configuration is used.

Value

Integer vector of set indices for the matching gene sets. Each set index refers to a row in the data frame returned by fetchAllSets.

Author

Aaron Lun

Examples

out <- searchSetText("9606", "cancer")
fetchSomeSets("9606", out[1])
#>                            name
#> 1 SOGA_COLORECTAL_CANCER_MYC_DN
#>                                                                          description
#> 1 http://www.gsea-msigdb.org/gsea/msigdb/human/geneset/SOGA_COLORECTAL_CANCER_MYC_DN
#>   size collection number
#> 1   71          3      2

out <- searchSetText("9606", "innate immun*")
fetchSomeSets("9606", out[1])
#>                            name
#> 1 REACTOME_INNATE_IMMUNE_SYSTEM
#>                                                                          description
#> 1 http://www.gsea-msigdb.org/gsea/msigdb/human/geneset/REACTOME_INNATE_IMMUNE_SYSTEM
#>   size collection number
#> 1 1118          5    230