Validate Gesel database files

Validate Gesel database and gene mapping files against the specification at https://github.com/gesel-inc/gesel-spec.

validateDatabaseFiles(path, species, num.genes)

validateGeneFiles(path, species, types = NULL)

Arguments

path

String containing the path to a directory containing the database files or gene mapping files, for validateDatabaseFiles and validateGeneFiles respectively.

species

String specifying the species in the form of its NCBI taxonomy ID.

num.genes

Integer scalar specifying the total number of genes available for this species.

types

Character vector specifying the types of gene names to validate, e.g.,"symbol", "entrez", or "ensembl", If NULL, all detected files for species in path are checked.

Value

validateDatabaseFiles returns NULL invisibly.

validateGeneFiles returns the number of genes, to be used as num.genes.

In both functions, invalid formatting will cause an error to be raised.

Author

Aaron Lun

Examples

example(prepareDatabaseFiles, echo=FALSE)
validateDatabaseFiles(output, "9606", num.genes)