Fetch all gene set collections

Fetch information about all gene set collections in the Gesel database.

fetchAllCollections(species, config = NULL)

Arguments

species

String containing the NCBI taxonomy ID of the species of interest.

config

Configuration list, typically created by newConfig. If NULL, the default configuration is used.

Value

Data frame of gene set collection information. Each row represents a collection and contains:

title, string containing the title for the collection.

description, string containing a description for the collection.

maintainer, string containing the identity of the collection's maintainer.

source, string containing the source of origin of the collection.

start, integer containing the set index for the first gene set in this collection. The set index refers to a row in the data frame returned by fetchAllSets.

size, integer scalar specifying the number of gene sets in the collection.

Author

Aaron Lun

Examples

out <- fetchAllCollections("9606")
head(out)
#>                                                title
#> 1                                      Gene ontology
#> 2                   MSigDB positional gene sets (C1)
#> 3 MSigDB chemical and genetic perturbations (C2 CGP)
#> 4                 MSigDB canonical pathways (C2 PID)
#> 5            MSigDB canonical pathways (C2 Reactome)
#> 6        MSigDB canonical pathways (C2 WikiPathways)
#>                                                                                                                 description
#> 1 Gene sets defined from the Gene Ontology (version 2022-07-01), sourced from the Bioconductor package org.Hs.eg.db 3.16.0.
#> 2      Gene annotations for this collection are derived from the Chromosome and Karyotype band tracks from Ensembl BioMart.
#> 3                                     Gene sets that represent expression signatures of genetic and chemical perturbations.
#> 4                                                                  Gene sets derived from the Pathway Interaction Database.
#> 5                       Reactome gene sets are derived from Reactome and have been filtered to remove inter-set redundancy.
#> 6                       WikiPathways gene sets are retrieved directly from the latest release of the WikiPathways database.
#>   maintainer
#> 1  Aaron Lun
#> 2  Aaron Lun
#> 3  Aaron Lun
#> 4  Aaron Lun
#> 5  Aaron Lun
#> 6  Aaron Lun
#>                                                                         source
#> 1 https://github.com/LTLA/gesel-feedstock/blob/gene-ontology-v1.0.0/go/build.R
#> 2           https://github.com/LTLA/gesel-feedstock/blob/master/msigdb/build.R
#> 3           https://github.com/LTLA/gesel-feedstock/blob/master/msigdb/build.R
#> 4           https://github.com/LTLA/gesel-feedstock/blob/master/msigdb/build.R
#> 5           https://github.com/LTLA/gesel-feedstock/blob/master/msigdb/build.R
#> 6           https://github.com/LTLA/gesel-feedstock/blob/master/msigdb/build.R
#>   start  size
#> 1     1 18933
#> 2 18934   300
#> 3 19234  3405
#> 4 22639   196
#> 5 22835  1654
#> 6 24489   733